i dont remember where i was at
