+++ public/HTML/signup.html
@@ -102,6 +102,7 @@
             <button type="submit">Signup</button>
         </form>
     </div>
+    
 
 </body>
 </html>
