+++ routes/api.js
@@ -1,19 +1,18 @@
 const express = require("express");
 const router = express.Router();
-const User = require("../models/user");
+const User = require("../models/accounts");
 const Menu = require("../models/menuItem");
-const Employee = require("../models/employee");
 
-// Add a new user
-router.post("/users", async (req, res) => {
-    try {
-        const user = new User(req.body);
-        await user.save();
-        res.status(201).json(user);
-    } catch (error) {
-        res.status(400).json({ error: error.message });
-    }
-});
+// Add a new user - this is being reworked idk if its needed
+// router.post("/account", async (req, res) => {
+//     try {
+//         const user = new User(req.body);
+//         await user.save();
+//         res.status(201).json(user);
+//     } catch (error) {
+//         res.status(400).json({ error: error.message });
+//     }
+// });
 
 // Add a new menu item
 router.post("/menu", async (req, res) => {
@@ -21,17 +20,6 @@ router.post("/menu", async (req, res) => {
         const menuItem = new Menu(req.body);
         await menuItem.save();
         res.status(201).json(menuItem);
-    } catch (error) {
-        res.status(400).json({ error: error.message });
-    }
-});
-
-// Add a new employee
-router.post("/employees", async (req, res) => {
-    try {
-        const employee = new Employee(req.body);
-        await employee.save();
-        res.status(201).json(employee);
     } catch (error) {
         res.status(400).json({ error: error.message });
     }
