+++ routes/index.js
@@ -57,6 +57,4 @@ router.get('/addItem', (req, res) => {
     res.sendFile(path.join(__dirname, '../public/HTML/addMenuItem.html'));
 });
 
-
-
 module.exports = router;
